pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "altman-zscore"
version = "0.1.0"
description = "Calculate Altman Z-Score for companies using SEC EDGAR data"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.0",
    "yfinance>=0.2.60",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "sec-edgar-downloader>=5.0.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "requests-mock>=1.11.0",
]
dev = [
    "types-requests>=2.31.0",
    "pandas-stubs>=2.2.0",
    "types-python-dateutil>=2.8.2",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["altman_zscore"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=altman_zscore"